// src/types/composer.ts
export enum CModule {
  CONNECT = 'CONNECT',
  CONVENE = 'CONVENE',
  COLLABORATE = 'COLLABORATE',
  CONTRIBUTE = 'CONTRIBUTE',
  CONVEY = 'CONVEY',
}